requirements.txt
# Core dependencies for Predictive Gift Selection System
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
numpy>=1.24.0
apscheduler>=3.10.0
python-multipart>=0.0.6
httpx>=0.24.0
gender-guesser>=0.4.0
catboost>=1.2.0
optuna>=3.0.0
click>=8.0.0
tabulate>=0.9.0
python-dotenv>=0.21.0 # For explicit .env file loading

# Production server
gunicorn>=20.1.0

# PostgreSQL support for Heroku
psycopg2-binary>=2.9.5